chore: Update ESLint configuration and add dropdown functionality
The ESLint configuration has been updated to extend the "airbnb-base" rules and include the "browser" and "node" environments.

Additionally, a new function `arrayToDropdown` has been added to convert an array of items into a dropdown menu. The function is documented with JSDoc comments and returns a div containing all the items as a dropdown menu. The CSS styles for the dropdown have also been added in a separate file.